+++ admin/admin-hooks.php
@@ -4,7 +4,7 @@
 */
 class H_Instafeed_Admin_Hooks {
   function __construct() {
-    add_action( 'admin_enqueue_scripts', array($this, 'admin_enqueue') );
+    // add_action( 'admin_enqueue_scripts', array($this, 'admin_enqueue') );
   }
 
   /*
@@ -12,9 +12,9 @@ class H_Instafeed_Admin_Hooks {
   */
   function admin_enqueue() {
     // Stylesheet
-    wp_enqueue_style( 'app-admin' , PLUGIN_NAME_DIR . 'assets/css/app-admin.css' );
+    wp_enqueue_style( 'app-admin' , H_INSTAFEED_URL . '/assets/css/app-admin.css' );
 
     // Javascript
-    wp_enqueue_script( 'app-admin', PLUGIN_NAME_DIR . 'assets/js/app-admin.js', false, false, true );
+    wp_enqueue_script( 'app-admin', H_INSTAFEED_URL . '/assets/js/app-admin.js', false, false, true );
   }
 }