+++ h-instagram-feed.php
@@ -1,17 +1,19 @@
 <?php
 /*
 Plugin Name: Edje Instagram Feed
-Description: Display Instagram photos from any user in WordPress theme. Works without access token! 
+Description: Display Instagram photos from any user in WordPress theme. Works without access token! SHORTCODE: <code>[instafeed your.username]</code>
 Plugin URI: https://github.com/hrsetyono/wp-instagram-feed
-Author: Henner Setyono
-Author URI: http://pixelstudio.id
-Version: 1.0.0
+Author: Pixel Studio ID
+Author URI: https://pixelstudio.id
+Version: 0.2.0
 */
 
 // If this file is called directly, abort.
 if( !defined( 'WPINC' ) ) { die; }
 define( 'H_INSTAFEED_VERSION', '0.1.0' );
-define( 'H_INSTAFEED_DIR', plugins_url( '', __FILE__ ) );
+define( 'H_INSTAFEED_URL', plugins_url( '', __FILE__ ) );
+define( 'H_INSTAFEED_PATH', untrailingslashit( plugin_dir_path( __FILE__ ) ) );
+
 
 
 new H_Instafeed();
@@ -34,7 +36,7 @@ class H_Instafeed {
 		Load the required dependencies for Admin pages.
 	*/
 	function admin_load() {
-		require_once PLUGIN_NAME_DIR . 'admin/admin-hooks.php';
+		require_once H_INSTAFEED_PATH . '/admin/admin-hooks.php';
 
 		new H_Instafeed_Admin_Hooks();
 	}
@@ -43,7 +45,7 @@ class H_Instafeed {
 		Load the required dependencies for Public pages.
 	*/
 	function public_load() {
-		require_once PLUGIN_NAME_DIR . 'public/public-hooks.php';
+		require_once H_INSTAFEED_PATH . '/public/public-hooks.php';
 
 		new H_Instafeed_Public_Hooks();
 	}
@@ -64,7 +66,10 @@ class H_Instafeed {
 
 	///
 
-	static function get(  ) {
-
+	/*
+    Alias for get_photos()
+  */
+	static function get( $username, $endcursor = '' ) {
+		return H_Instafeed_Public_Hooks::get_data( $username, $endcursor );
 	}
 }