+++ public/public-hooks.php
@@ -4,33 +4,45 @@
 */
 class H_Instafeed_Public_Hooks {
   function __construct() {
-    add_shortcode( 'instafeed', array($this, 'instafeed') );
-    add_action( 'enqueue_scripts', array($this, 'public_enqueue') );
+    add_shortcode( 'instafeed', array($this, 'instafeed_shortcode') );
+    add_action( 'wp_enqueue_scripts', array($this, 'register_assets') );
   }
 
   /*
     Display instagram slider of a specific username
 
-    @shortcode [instafeed USERNAME]
+    @shortcode [instafeed]
   */
-  function instafeed( $atts, $content = null ) {
-    if( !isset($atts[0]) ) { return false; }
+  function instafeed_shortcode( $atts, $content = null ) {
+    if( !isset($atts['username']) ) { return false; } // must have username
 
-    $username = $atts[0];
-    $feed = $this->get_photos( $username );
-    return $feed;
+    $this->enqueue_assets();
+
+    // filter attributes and set default value
+    $atts = shortcode_atts( array(
+      'username' => '',
+      'items-per-slide' => '4/3/2',
+    ), $atts );
+
+    $this->output_slider( $atts ); 
   }
   
 
   /*
     @action enqueue_scripts
   */
-  function public_enqueue() {
-    // Stylesheet
-    wp_enqueue_style( 'app-public' , PLUGIN_NAME_DIR . 'assets/css/app-public.css' );
+  function register_assets() {
+    if( !wp_script_is( 'h-slider', 'enqueued' ) ) {
+      wp_register_style( 'h-slider' , H_INSTAFEED_URL . '/assets/css/h-slider.css' );
+    }
+
+    if( !wp_script_is( 'h-lightbox', 'enqueued' ) ) {
+      wp_register_style( 'h-lightbox' , H_INSTAFEED_URL . '/assets/css/h-lightbox.css' );
+    }
 
     // Javascript
-    wp_enqueue_script( 'app-public', PLUGIN_NAME_DIR . 'assets/js/app-public.js', false, false, true );
+    wp_register_script( 'h-slider', H_INSTAFEED_URL . '/assets/js-vendor/h-slider.min.js', array('jquery'), false, true );
+    wp_register_script( 'h-lightbox', H_INSTAFEED_URL . '/assets/js-vendor/h-lightbox.min.js', array('jquery'), false, true );
   }
 
 
@@ -44,8 +56,10 @@ class H_Instafeed_Public_Hooks {
     @param $endcursor (string) - Bit from the JSON data used to get the next 20 photos (pagination)
     @return array
   */
-  private function get_photos( $username, $endcursor = '' ) {
-    $cache = new Instagram\Storage\CacheManager('../../uploads/instafeed-cache');
+  static function get_data( $username, $endcursor = '' ) {
+    require_once H_INSTAFEED_PATH . '/vendor/autoload.php';
+    
+    $cache = new Instagram\Storage\CacheManager('../../h-instafeed-cache');
     $api   = new Instagram\Api( $cache );
 
     $api->setUserName( $username );
@@ -57,5 +71,42 @@ class H_Instafeed_Public_Hooks {
 
     $feed = $api->getFeed();
     return $feed;
+  }
+
+  /*
+    Enqueue necessary assets
+  */
+  private function enqueue_assets() {
+    if( !wp_script_is( 'h-slider', 'enqueued' ) ) {
+      wp_enqueue_style( 'h-slider' );
+    }
+
+    if( !wp_script_is( 'h-lightbox', 'enqueued' ) ) {
+      wp_enqueue_style( 'h-lightbox' );
+    }
+
+    wp_enqueue_script( 'h-slider' );
+    wp_enqueue_script( 'h-lightbox' );
+  }
+
+  /*
+    Slider template
+  */
+  private function output_slider( $atts ) {
+    $data = self::get_data( $atts['username'] );
+
+    ?>
+    <div class="h-instafeed-slider" data-ips="<?php echo $atts['items-per-slide']; ?>">
+      <?php foreach( $data->medias as $photo ):
+        $caption = preg_replace( '/\s#\w+/', '', $photo->caption );
+        $src = $photo->thumbnails[3]->src; ?>
+        
+        <a href="<?php echo $photo->displaySrc; ?>" title="<?php echo $caption; ?>">
+          <img src="<?php echo $src; ?>">
+        </a>
+        
+      <?php endforeach; ?>
+    </div>
+    <?php
   }
 }