+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+# 0.2.0
 
-- Starting from [boilerplate](https://github.com/hrsetyono/wp-plugin-boilerplate)
+- Updated hSlider to v 2.1.1
+- Added attribute `items-per-slide` to shortcode.
+- Added attribute `username` to shortcode.
+
+# 0.1.0
+
+- Added shortcode `[instafeed]`.
+- Added function to get JSON data `H_Instafeed::get()`.